refactor(mfa): split Context module to break import cycle
OutcomeScreenBase and AuthenticationMethodDescription pulled the full
Context module (Provider -> Main -> State -> stateReducer -> config ->
scenarios), forming an import cycle that resolved to
customConfig(undefined) at module-load and crashed the
sortTransactionsPending3DSReview test suite at load.

- Extract the two React Context objects + their consumer hooks into
  leaf files MultifactorAuthenticationStateContext.ts and
  MultifactorAuthenticationActionsContext.ts (no downstream imports).
- Merge State.tsx into Provider.tsx and rename to
  MultifactorAuthenticationComposedContextProviders.tsx; drop
  ComposeProviders for direct nesting.
- Rename Main.tsx to MultifactorAuthenticationMainContext.tsx.
- Point UI primitives at the leaf context files directly so their
  import graph stops at React + types.
