+++ src/components/MultifactorAuthentication/Context/index.ts
@@ -1,6 +1,5 @@
-export {default as MultifactorAuthenticationContextProviders} from './Provider';
-export {useMultifactorAuthentication} from './Main';
-
-export {useMultifactorAuthenticationState, useMultifactorAuthenticationActions} from './State';
-
+export {default as MultifactorAuthenticationContextProviders} from './MultifactorAuthenticationComposedContextProviders';
+export {useMultifactorAuthentication} from './MultifactorAuthenticationMainContext';
+export {useMultifactorAuthenticationActions} from './MultifactorAuthenticationActionsContext';
+export {useMultifactorAuthenticationState} from './MultifactorAuthenticationStateContext';
 export {default as usePromptContent} from './usePromptContent';
